requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
pip~=23.3.1
pytz~=2023.3.post1
asgiref~=3.7.2
sqlparse~=0.4.4
setuptools~=63.2.0
PyJWT~=2.8.0
djangorestframework-simplejwt~=5.3.0
django-cors-headers~=4.3.1
drf-yasg~=1.21.7